refactor(icp): move the native project loader out of the library
Parsing a manifest is a frontend concern: it means walking the
filesystem, reading YAML, globbing canister directories and resolving
file references. `Project`, `ProjectLoad` and `ProjectRootLocate` stay in
`icp` — a consumer that deploys must still be able to receive a project —
but the native, filesystem-touching implementations behind those ports
now live in `icp-cli`:

- `icp::manifest::{ProjectRootLocateImpl, load_manifest_from_path}`
  -> `icp_cli::manifest`
- `icp::project` (manifest consolidation) and `icp::{ProjectLoadImpl, Lazy}`
  -> `icp_cli::project`

The manifest *shapes* stay in `icp`: they are plain serde types with no
filesystem in them, `BuildSteps`/`RootKeySpec` and friends are part of
the `Project` model the builder consumes, and `schema-gen` generates the
published JSON schemas from them. Their modules are now `pub` so the
loader can reach the same types it always did.

`ProjectLoadError`'s two loader-specific variants collapse into one boxed
`Load` variant, so the library's port does not name any one loader's
error type. It is `#[snafu(transparent)]`, and the CLI loader keeps
variants with the same messages, so error output is unchanged:
`failed to load project` / `failed to load project manifest` /
`failed to parse manifest at '<path>'`. `Locate` stays a real variant —
callers match on it to detect "no project here".

`icp::context::initialize` now receives the directories, the root
locator and the project loader instead of constructing them, since the
loader (and the `$PWD`-vs-getcwd resolution that feeds it) is the
frontend's to build. `DEFAULT_LOCAL_NETWORK_{BIND,PORT}` move to
`icp::network`, where the default network they describe is defined.
