+++ crates/icp-cli/Cargo.toml
@@ -34,6 +34,7 @@ dunce.workspace = true
 elliptic-curve.workspace = true
 flate2.workspace = true
 futures.workspace = true
+glob.workspace = true
 tar.workspace = true
 hex.workspace = true
 httptest.workspace = true
@@ -46,6 +47,7 @@ icp-canister-interfaces.workspace = true
 icp-events.workspace = true
 icp = { workspace = true, features = ["clap"] }
 icrc-ledger-types.workspace = true
+indexmap.workspace = true
 indicatif.workspace = true
 indoc.workspace = true
 itertools.workspace = true
@@ -56,6 +58,7 @@ num-integer.workspace = true
 num-traits.workspace = true
 open.workspace = true
 p256.workspace = true
+pathdiff.workspace = true
 pem.workspace = true
 phf.workspace = true
 pkcs8.workspace = true
